backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9
python-dotenv>=1.0
pytest>=8.0
pytest-django>=4.8
black>=24.0
isort>=5.13
ruff>=0.4
